move test so it can use IPFS with no conflict
